// src/utils/enum.ts
export enum FABRIC_NAME {
  BAG_VOILE = "BAG VOILE",
  WISTERIA_VOLIE = "WISTERIA VOLIE",
  SHEETING = "SHEETING",
  WINDOWPANE_GAUZ = "WINDOWPANE GAUZ",
  FP001 = "FP001",
  WAFFEL = "WAFFEL",
  RIB = "2X2 RIB",
  FRENCH_TERRY = "FRENCH TERRY",
  VISCOUS_LYCRA = "VISCOUS LYCRA",
  OTHER_KNITS = "OTHER KNITS",
  LACES = "LACES",
  SHIFILLI = "SHIFILLI",
  SPECIAL_FABRICS = "SPECIAL FABRICS",
  IMPORTED_FABRICS = "IMPORTED FABRICS",
  CHINA_LACE = "CHINA LACE",
}

export enum PROCESSES {
  DYING = "DYING",
  PRINTING = "PRINTING",
  MOCK_UP = "MOCK UP",
  LACES = "LACES",
  SHIFFLY = "SHIFFLY",
  WASHING = "WASHING",
}

export enum SKIPPED_STAGES {
  PURCHASE = "PURCHASE",
  SUBMISSION = "SUBMISSION",
  FOB = "FOB",
  BULK = "BULK",
  FABRIC_AUDIT = "FABRIC AUDIT",
  PRODUCTION = "PRODUCTION",
  TOP = "TOP",
  WEB = "WEB",
  SIZE_SET = "SIZE SET",
}

export enum TRIMS {
  LABEL_COPPER = "LABEL COPPER",
  TAG = "TAG",
  STICKER = "STICKER",
  POLY_BAG = "POLY BAG",
  STITCHING_THREAD = "STITCHING THREAD",
}

export enum ACCESSORIES {
  WASHCARE_LABEL = "WASHCARE LABEL",
  PRICE_TAG = "PRICE TAG",
  SIZE_LABEL = "SIZE LABEL",
  BUTTON = "BUTTON",
  ZIP = "ZIP",
  HOOK = "HOOK",
}